+++ src/components/List/ListItem.svelte
@@ -1,6 +1,7 @@
 <script>
   import { createEventDispatcher } from "svelte";
   import Icon from "../Icon";
+  import createRipple from "../Ripple/ripple.js";
 
   export let icon = "";
   export let id = "";
@@ -13,10 +14,11 @@
   export let to = "";
   export let selected = false;
   export let tabindex = null;
-  export let basicClasses = "my-1 ripple-gray hover:bg-gray-trans transition p-4 cursor-pointer text-gray-700 flex items-center z-10";
+  export let basicClasses = "relative overflow-hidden my-1 transition p-4 cursor-pointer text-gray-700 flex items-center z-10";
   export let itemClasses = "";
   export let selectedClasses = "bg-gray-200";
 
+  const ripple = createRipple();
   const dispatch = createEventDispatcher();
 
   function change() {
@@ -41,8 +43,10 @@
 </style>
 
 <li
+  use:ripple
   class="{basicClasses} {selected ? selectedClasses : ''}"
   class:navigation
+  class:hover:bg-gray-transLight={!navigation}
   class:py-2={dense}
   class:text-gray-600={disabled}
   {tabindex}