+++ src/components/RadioButton/variants.js
@@ -1,3 +1,3 @@
 module.exports = {
-  all: color => [`ripple-${color}-500`, `ripple-${color}`, `text-${color}-500`]
+  all: color => [`text-${color}-500`]
 };