+++ src/components/Ripple/Ripple.svelte
@@ -1,5 +1,10 @@
 <script>
   export let color = "primary";
+  export let noHover = false;
+  import createRipple from "../Ripple/ripple.js";
+
+  $: ripple = createRipple(color, true);
+  $: hoverClass = `hover:bg-${color}-transDark`;
 </script>
 
 <style>
@@ -8,6 +13,8 @@
   }
 </style>
 
-<span class="z-40 p-2 rounded-full flex items-center justify-center top-0 left-0 ripple-{color} hover:bg-{color}-trans">
+<span
+  use:ripple
+  class="z-40 p-2 rounded-full flex items-center justify-center top-0 left-0 {noHover ? '' : hoverClass}">
   <slot />
 </span>