+++ src/components/Ripple/variants.js
@@ -1,8 +1,3 @@
 module.exports = {
-  all: color => [
-    `ripple-${color}-500`,
-    `ripple-${color}`,
-    `hover:bg-${color}-trans`,
-    `bg-${color}-trans`
-  ]
+  all: color => [`hover:bg-${color}-trans`, `bg-${color}-trans`]
 };