+++ src/components/Switch/variants.js
@@ -1,8 +1,3 @@
 module.exports = {
-  all: color => [
-    `ripple-${color}-500`,
-    `ripple-${color}`,
-    `bg-${color}-200`,
-    `bg-${color}-400`
-  ]
+  all: color => [`bg-${color}-200`, `bg-${color}-400`]
 };