+++ src/components/Tabs/variants.js
@@ -1,3 +1,3 @@
 module.exports = {
-  all: color => [`bg-${color}-50`, `ripple-${color}-500`, `bg-${color}-700`]
+  all: color => [`bg-${color}-50`, `bg-${color}-700`]
 };