+++ src/utils/classes.js
@@ -3,7 +3,6 @@ export default function utils(color, defaultDepth = 500) {
     bg: depth => `bg-${color}-${depth || defaultDepth} `,
     border: depth => `border-${color}-${depth || defaultDepth} `,
     txt: depth => `text-${color}-${depth || defaultDepth} `,
-    ripple: depth => `ripple-${color}-${depth || defaultDepth} `,
     caret: depth => `caret-${color}-${depth || defaultDepth} `
   };
 }