+++ src/utils/style.js
@@ -1,33 +1,7 @@
 module.exports = {
-  rippleAfter: {
-    content: "''",
-    display: "block",
-    position: "absolute",
-    width: "100%",
-    height: "100%",
-    top: 0,
-    left: 0,
-    pointerEvents: "none",
-    backgroundRepeat: "no-repeat",
-    backgroundPosition: "50%",
-    transform: "scale(10, 10)",
-    opacity: 0,
-    transition: "transform .4s ease-in, opacity 0.8s ease-out"
-  },
-  ripple: {
-    position: "relative",
-    overflow: "hidden",
-    transform: "translate3d(0, 0, 0)"
-  },
-  rippleActiveAfter: {
-    transform: "scale(0, 0)",
-    opacity: 0.3,
-    transition: "0s"
-  },
   addUtility({
     prop = "caret-color",
     className = ".caret",
-    value = a => a,
     defaultVariant = 500
   }) {
     return function({ e, addUtilities, theme }) {